# Consolidate redundant backend creation in tests