# Purpose

Creating backend object every test cases is redundant. We can make test faster by eliminating this redundant object creation.

# Result

Before

```
Chrome 65.0.3325 (Mac OS X 10.12.6): Executed 4089 of 4089 SUCCESS (2 mins 18.853 secs / 2 mins 17.784 secs)
Firefox 58.0.0 (Mac OS X 10.12.0): Executed 4089 of 4089 SUCCESS (9 mins 7.102 secs / 8 mins 25.818 secs)
```

After

```
Chrome 65.0.3325 (Mac OS X 10.12.6): Executed 4089 of 4089 SUCCESS (1 min 15.038 secs / 1 min 1.975 secs)
Firefox 59.0.0 (Mac OS X 10.12.0): Executed 4089 of 4089 SUCCESS (5 mins 6.947 secs / 4 mins 25.21 secs)
```

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/pair-code/deeplearnjs/861)
<!-- Reviewable:end -->