@Lewuathe 2018-04-05
@dsmilkov @nsthorat Sorry bothering you. But could you take a look?
Since it's not the ops update, we can merge this even in API freeze, I think.