@dsmilkov 2018-04-05
<img class="emoji" title=":lgtm_strong:" alt=":lgtm_strong:" align="absmiddle" src="https://reviewable.io/lgtm_strong.png" height="20" width="61"/> Thank you for this *amazing* contribution!

---

Reviewed 1 of 1 files at r2.
Review status: all files reviewed at latest revision, 1 unresolved discussion.

---



*Comments from [Reviewable](https://beta.reviewable.io/reviews/tensorflow/tfjs-core/861#-:-L9Lq_xZ-ZqAC4rSPPTA:b-gk86mx)*
<!-- Sent from Reviewable.io -->